% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrZ.R
\name{corrZ}
\alias{corrZ}
\title{Correlation Between Two Correlated Logrank Tests for Time-To-Event Endpoints}
\usage{
corrZ(
  e.strat = list(AandB1 = 120, AnotB1 = 100, AandB2 = 120, BnotA2 = 80),
  e.unstr = list(A1 = 220, B2 = 200, minAorB = 300),
  r.strat = list(AandB = 1/2, AnotB = 1/2, BnotA = 1/2),
  r.unstr = list(A = 1/2, B = 1/2, AandB = 1/2),
  pAandB.unstr = 0.5,
  strat = c("N", "Y"),
  method = c("Approximation", "Observed")
)
}
\arguments{
\item{e.strat}{Numbers of events for stratified analysis
e.strat = list(AandB1, AnotB1, AandB2, BnotA2)
AandB1: events for AnB at tau
AnotB1: events for A and not B at tau
AandB2: events for A and B at tau'
BnotA2: events for B and not A at tau'}

\item{e.unstr}{Numbers of events for unstratified analysis
e.unstrat = list(A1, B2, minAorB).
A1: events for A at tau
B2: events for B at tau'
minAorB: events for AUB at min(tau, tau')}

\item{r.strat}{Proportions of experimental subjects for stratified
analysis in each of subgroup:
r.strat = list(AandB = 1/2, AnotB=0, BnotA=1/2).}

\item{r.unstr}{Proportions of experimental subjects in A, B and AandB
r.unstr = list(A=1/2, B=1/2, AandB=1/2). For simplicity,
no consideration of r change over time.}

\item{pAandB.unstr}{Proportion of subjects included in both A and B among all subjects in A or B.}

\item{strat}{stratified analysis flag, "Y" or "N". Default, "Y". The stratified analysis
means that testing HA is stratified by B, and vice versa. e.unstr, r.unstr, and
gamma.unstr are required for unstratified analysis;
e.strat and r.strat are required for stratified analysis.}

\item{method}{Method for randomization ratio, i.e., proportion of subjects in experimental arm.
"Approximation" or "Observed". "Approximation" method assumes the same randomization ratio in every stratum;
"Observed" method estimates the randomization ratio by the data.
At study design stage without data, "Approximation" method should be used.}
}
\value{
corr: Correlation of the two logrank test statistics

@references
}
\description{
This function calculates the correlation between two correlated logrank tests
Z_A(tau) and Z_B(tau') at analysis times tau and tau'.
The correlation is due to the overlapping populations and timing.
#'
}
\examples{
   #Example 2 in the manuscript.
   #1. Correlation between H1 and H2 at tau -- unstratified with approx.
   corrZ(e.strat = list(AandB1=120, AnotB1=100, AandB2=120, BnotA2=80),
   e.unstr = list(A1=220, B2=200, minAorB=300),                 
   r.strat = list(AandB = 1/2, AnotB=1/2, BnotA=1/2),
   r.unstr = list(A=1/2, B=1/2, AandB=1/2), pAandB.unstr = 0.5,
   strat = c("N"), method = c("Approximation"))
   
   #2. Correlation between H1 and H2  at tau -- unstratified Observed randomization ratios
   corrZ(e.strat = list(AandB1=120, AnotB1=100, AandB2=120, BnotA2=80),
   e.unstr = list(A1=220, B2=200, minAorB=300),                 
   r.strat = list(AandB = 1/2, AnotB=1/2, BnotA=1/2),
   r.unstr = list(A=1/2, B=1/2, AandB=1/2), pAandB.unstr = 0.5,
   strat = c("N"), method = c("Observed"))
   
   #3. Correlation between H1 and H1  at tau -- unstratified Observed randomization ratios
   corrZ(e.strat = list(AandB1=120, AnotB1=100, AandB2=120, BnotA2=80),
   e.unstr = list(A1=220, B2=220, minAorB=220),                 
   r.strat = list(AandB = 1/2, AnotB=1/2, BnotA=1/2),
   r.unstr = list(A=1/2, B=1/2, AandB=1/2), pAandB.unstr = 1,
   strat = c("N"), method = c("Approximation"))
   
   #4. Correlation between H1 and H1 at different times tau and tau' -- unstratified Observed randomization ratios
   corrZ(e.strat = list(AandB1=120, AnotB1=100, AandB2=120, BnotA2=80),
   e.unstr = list(A1=220, B2=280, minAorB=220),                 
   r.strat = list(AandB = 1/2, AnotB=1/2, BnotA=1/2),
   r.unstr = list(A=1/2, B=1/2, AandB=1/2), pAandB.unstr = 1,
   strat = c("N"), method = c("Approximation"))
   
   #5. Correlation between H1 and H2 at tau -- stratified with approx.
   corrZ(e.strat = list(AandB1=120, AnotB1=100, AandB2=120, BnotA2=80),
   e.unstr = list(A1=220, B2=200, minAorB=300),                 
   r.strat = list(AandB = 1/2, AnotB=1/2, BnotA=1/2),
   r.unstr = list(A=1/2, B=1/2, AandB=1/2), pAandB.unstr = 0.5,
   strat = c("Y"), method = c("Approximation"))
   
   #6. Correlation between H1 and H2 at tau -- stratified Observed randomization ratios
   corrZ(e.strat = list(AandB1=120, AnotB1=100, AandB2=120, BnotA2=80),
   e.unstr = list(A1=220, B2=200, minAorB=300),                 
   r.strat = list(AandB = 1/2, AnotB=1/2, BnotA=1/2),
   r.unstr = list(A=1/2, B=1/2, AandB=1/2), pAandB.unstr = 0.5,
   strat = c("Y"), method = c("Observed"))
   
   #7. Correlation between H1 and H3 at tau -- stratified Observed randomization ratios
   corrZ(e.strat = list(AandB1=220, AnotB1=0, AandB2=220, BnotA2=280),
   e.unstr = list(A1=220, B2=500, minAorB=500),                 
   r.strat = list(AandB = 1/2, AnotB=1/2, BnotA=1/2),
   r.unstr = list(A=1/2, B=1/2, AandB=1/2), pAandB.unstr = 0.5,
   strat = c("Y"), method = c("Observed"))    
   
   #8. Correlation between H2 and H3 at tau -- stratified Observed randomization ratios
   corrZ(e.strat = list(AandB1=200, AnotB1=0, AandB2=200, BnotA2=300),
   e.unstr = list(A1=200, B2=500, minAorB=500),                 
   r.strat = list(AandB = 1/2, AnotB=1/2, BnotA=1/2),
   r.unstr = list(A=1/2, B=1/2, AandB=1/2), pAandB.unstr = 0.5,
   strat = c("Y"), method = c("Observed"))    
   
}
